requirements.txt
alembic==1.0.0
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
cffi==1.11.5
click==6.7
cryptography==2.3.1
decorator==4.3.0
dominate==2.3.1
Flask==1.0.2
flask-blueprint==1.2.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.4
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
PyMySQL==0.9.2
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.10
traitlets==4.3.2
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
